Deduplicate module
When bt runs a build script that imports bt, bt is imported with
a different name so it runs again.
With this patch, the main code is in `__init__.py` which is imported by
`__main__.py` with the name `bt` so that `import bt` in the build script
reuses it. Thus module duplication is eliminated.

An obscure consequence of duplicating the main module when bt is
executed directly is that pickled instances of bt classes belong to
different modules which are included in the cache file. The module
name in the cache file thus differs depending on how bt is run. When bt
runs by being imported, its module name is `bt`; when it is executed,
its name is `__main__` and `import bt` in `bs` loads a new instance of
`__main__` named `bt`. If bt is run by `./bs`, then the cache contains
a reference to `bt` but `__main__` is the module that performs the build
logic so unpickling the cache produces objects from the unused module
which—in addition to the obvious resulting problems—breaks equality
checks and thus invalidates the cache.