Raise `KeyboardInterrupt` only in live threads
Apparently `KeyboardInterrupt` is always raised in the main thread.
If it's raised while the main thread is dead, no exception handler is
called so the exception is logged.
This commit suppresses the message by trapping `SIGINT` and raising
`KeyboardInterrupt` in every live thread instead.

This commit also introduces a `threading.excepthook` in order to catch
`KeyboardInterrupt`s in non-main threads.